server/requirements.txt
flask
python-binance
python-dotenv